+++ secretsmanager/resource_address.go
@@ -54,6 +54,8 @@ func resourceAddress() *schema.Resource {
 			// fields[]
 			"address":  schemaAddressField(),
 			"file_ref": schemaFileRefField(),
+			// custom[]
+			"custom": schemaCustomField(),
 		},
 	}
 }
@@ -103,6 +105,14 @@ func resourceAddressCreate(ctx context.Context, d *schema.ResourceData, m interf
 			if err := SetFieldTypeInSchema(d, "file_ref", "fileRef"); err != nil {
 				return diag.FromErr(err)
 			}
+		}
+	}
+
+	if customData := d.Get("custom"); customData != nil {
+		if fields, err := customFieldsFromSchema(customData.([]interface{})); err != nil {
+			return diag.FromErr(err)
+		} else {
+			nrc.Custom = fields
 		}
 	}
 
@@ -194,6 +204,11 @@ func resourceAddressRead(ctx context.Context, d *schema.ResourceData, m interfac
 		return diag.FromErr(err)
 	}
 
+	customItems := getFieldItemsData(secret.RecordDict, "custom")
+	if err := d.Set("custom", customItems); err != nil {
+		return diag.FromErr(err)
+	}
+
 	d.SetId(uid)
 	return diags
 }
@@ -237,6 +252,15 @@ func resourceAddressUpdate(ctx context.Context, d *schema.ResourceData, m interf
 		if _, err := ApplyFieldChange("fields", "file_ref", d, secret); err != nil {
 			return diag.FromErr(err)
 		}
+	}
+
+	if d.HasChange("custom") {
+		customData := d.Get("custom").([]interface{})
+		fields, err := customFieldsFromSchema(customData)
+		if err != nil {
+			return diag.FromErr(err)
+		}
+		secret.RecordDict["custom"] = customFieldsToDict(fields)
 	}
 
 	secret.RawJson = core.DictToJson(secret.RecordDict)