+++ secretsmanager/resource_encrypted_notes_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/keeper-security/secrets-manager-go/core"
 )
 
+
 func TestAccResourceEncryptedNotes_create(t *testing.T) {
 	secretType := "encryptedNotes"
 	secretFolderUid := testAcc.getTestFolder()
@@ -194,6 +195,42 @@ func TestAccResourceEncryptedNotes_import(t *testing.T) {
 				ResourceName:      resourceName,
 				ImportState:       true,
 				ImportStateVerify: true,
+			},
+		},
+	})
+}
+
+func TestAccResourceEncryptedNotes_customField(t *testing.T) {
+	secretFolderUid := testAcc.getTestFolder()
+	secretUid := core.GenerateUid()
+	secretTitle := "tf_acc_encrypted_notes_custom"
+
+	config := fmt.Sprintf(`
+		resource "secretsmanager_encrypted_notes" "custom" {
+			folder_uid = "%v"
+			uid        = "%v"
+			title      = "%v"
+
+			custom {
+				type  = "text"
+				label = "Classification"
+				value = "confidential"
+			}
+		}
+	`, secretFolderUid, secretUid, secretTitle)
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		PreCheck:                 testAccPreCheck(t),
+		Steps: []resource.TestStep{
+			{
+				Config: config,
+				Check: resource.ComposeTestCheckFunc(
+					checkSecretExistsRemotely(secretUid),
+					resource.TestCheckResourceAttr("secretsmanager_encrypted_notes.custom", "custom.#", "1"),
+					resource.TestCheckResourceAttr("secretsmanager_encrypted_notes.custom", "custom.0.label", "Classification"),
+					resource.TestCheckResourceAttr("secretsmanager_encrypted_notes.custom", "custom.0.value", "confidential"),
+				),
 			},
 		},
 	})